split timelineSection
